+++ vulcanodb-test/src/test/java/es/nachobrito/vulcanodb/SentenceEmbeddingTest.java
@@ -66,11 +66,11 @@ public class SentenceEmbeddingTest {
 
     }
 
-    public static double[] convertFloatsToDoubles(float[] input) {
+    public static float[] convertFloatsToDoubles(float[] input) {
         if (input == null) {
             return null; // Or throw an exception - your choice
         }
-        double[] output = new double[input.length];
+        float[] output = new float[input.length];
         for (int i = 0; i < input.length; i++) {
             output[i] = input[i];
         }
